Develop

+++ app/config/settings.py
@@ -1,11 +1,10 @@
 from pydantic_settings import BaseSettings
-from decimal import Decimal
 
 
 class Settings(BaseSettings):
     symbol: str
-    open: Decimal
-    higher: Decimal
-    lower: Decimal
-    closed: Decimal
+    open: float
+    higher: float
+    lower: float
+    closed: float
     volume: int
